server/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
pdfplumber==0.9.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
uuid==1.30
python-dateutil==2.8.2